refactor: Move item into `PickOne` component
