+++ src/components/ModalPickOne/index.tsx
@@ -1,5 +1,5 @@
 import { Component } from './component';
-import { Item } from './Item';
+import { Item } from '../internal/PickOne/Item';
 
 export const ModalPickOne = Component as typeof Component & { Item: typeof Item };
